+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG
 ---------
 
 #### Next Release
+- Fixed issue with parallel verses loading (#104)
 
 #### v0.10.0 (2019-08-26)
 - Improved R8 configurations: